update: Accurate types, file setup

+++ internal/tui/model.go
@@ -1,15 +1,15 @@
 package tui
 
 type Model struct {
-	board  [19 * 19]int // 0 is empty, 1 is black, 2 is white
-	cursor [2]int       // Which cell is currently selected
-	moves  [][2]int     // List of moves made
-	turn   int          // 1 is black, 2 is white
+	board  [19 * 19]uint8 // 0 is empty, 1 is black, 2 is white
+	cursor [2]uint8       // Which cell is currently selected
+	moves  [][2]uint8     // List of moves made
+	turn   uint8          // 1 is black, 2 is white
 }
 
 func InitialModel() Model {
 	return Model{
-		board: [19 * 19]int{
+		board: [19 * 19]uint8{
 			1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 			1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
 			1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
@@ -30,8 +30,8 @@ func InitialModel() Model {
 			1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
 			1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 		},
-		cursor: [2]int{9, 9},
-		moves:  [][2]int{},
+		cursor: [2]uint8{9, 9},
+		moves:  [][2]uint8{},
 		turn:   1,
 	}
 }
